move PriceList to price
