+++ credmark/types/__init__.py
@@ -20,6 +20,6 @@ from .models.rpc import (
     RpcBlockRangeOutput
 )
 
-from .data.portfolio import (Portfolio, PriceList)
+from .data.portfolio import Portfolio
 from .data.position import Position, NativePosition, TokenPosition
-from .data.price import Price, TokenPairPrice
+from .data.price import Price, TokenPairPrice, PriceList
