+++ credmark/types/data/portfolio.py
@@ -1,12 +1,9 @@
 from typing import List
 from credmark.dto import (
-    DTO,
     DTOField,
     PrivateAttr,
     IterableListGenericDTO,
 )
-
-from .address import Address
 
 from .position import Position
 
@@ -18,15 +15,4 @@ class Portfolio(IterableListGenericDTO[Position]):
     class Config:
         schema_extra: dict = {
             'examples': [{'positions': [exp]} for exp in Position.Config.schema_extra['examples']]
-        }
-
-
-class PriceList(DTO):
-    prices: List[float]
-    tokenAddress: Address
-
-    class Config:
-        schema_extra: dict = {
-            'examples': [{'prices': [4.2, 2.3],
-                          'tokenAddress': '0x6B175474E89094C44Da98b954EedeAC495271d0F'}]
         }
